+++ contracts/pool_stable/src/lib.rs
@@ -1,5 +1,6 @@
 #![no_std]
 mod contract;
+mod error;
 mod math;
 mod storage;